Refactoring to 'wt debug'
